refactor: put server infrastructure in a new class
fix: warning related to (too) late monkey-patching

test: changed tests to new system
